Add a rm command and support for start/stop times. Move config to a file
Doing an 'ls' with the -v option will show time entry ids which can be
used with 'rm' to delete an entry. Start and stop now accept an
additional optional parameter to specify the time. Config has been
moved to a separate ~/.togglrc file, running a command without this file
will generate a default.