Add completed time entries w/ duration or end time
Changed the final argument for add time entry.

If the final argument begins with d its treated as a duration.
Otherwise, its treated like an end time

I reordered the add_time_entry flow to first create a 0 duration time entry and then append the start and end_time/durations as needed.

Note, if an end time and duration are passed the duration needs to be the _actual_ duration (the delta in seconds between start and stop.