TimeEntry.start(): use start endpoint when start_time is now
It happens, probably due to localization, that the current time is not
communicated properly.

In order start a time-entry with the current time we can use the
time_entries/start endpoint. The advantage is that we do not have
to handle start time which is managed by toggl web service.

https://github.com/toggl/toggl_api_docs/blob/master/chapters/time_entries.md

Signed-off-by: Federico Vaga <federico.vaga@gmail.com>