Allow Python 3.8
Arch Linux updated the system python version to 3.8, and I could not get this to run as it used to. After some digging I realized it was installing version 1.0!

I am using the current version from a local install on 3.8 right now, and, well, it has not exploded. I'll report back if I notice issues.